requirements.txt
fastapi
uvicorn
twilio
openai
python-dotenv
sqlalchemy
psycopg2-binary
python-multipart
pyngrok